requirements.txt
altgraph==0.17
apiai==1.2.3
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.3.1
auto-py-to-exe==2.8.0
autolike==1.0.1
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
cx-Freeze==6.5.3
distlib==0.3.1
Django==3.1.7
dnspython==2.1.0
download==0.3.5
Eel==0.12.4
filelock==3.0.12
future==0.18.2
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.0.0
httplib2==0.19.0
idna==2.10
importlib-metadata==3.10.0
MouseInfo==0.1.3
numpy==1.20.2
pefile==2019.4.18
Pillow==8.1.2
PyAutoGUI==0.9.52
pycparser==2.20
PyGetWindow==0.0.9
pyinstaller==4.2
pyinstaller-hooks-contrib==2021.1
pymongo==3.11.3
PyMsgBox==1.0.9
pynput==1.7.3
pyparsing==2.4.7
pyperclip==1.8.2
PyQt6==6.0.3
PyQt6-Qt6==6.0.2
PyQt6-sip==13.0.1
PyRect==0.1.4
PyScreeze==0.1.26
pyTelegramBotAPI==3.7.7
python-telegram-bot==13.2
pytube==10.4.1
PyTweening==1.0.3
pytz==2021.1
pywin32-ctypes==0.2.0
requests==2.25.1
selenium==3.141.0
six==1.15.0
sqlparse==0.4.1
tornado==6.1
tqdm==4.56.2
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.3
virtualenv==20.4.3
whichcraft==0.6.1
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.3.0
